refactor: enums 패키지 추가, enum 클래스 이동
