+++ src/main/java/http/HttpRequest.java
@@ -1,5 +1,7 @@
 package http;
 
+import http.enums.HttpMethod;
+
 import java.util.HashMap;
 import java.util.Map;
