+++ src/main/java/http/enums/HttpMethod.java
@@ -1,4 +1,4 @@
-package http;
+package http.enums;
 
 import java.util.HashMap;
 import java.util.Map;
